refactor: shrink 3 oversized api controllers and split activities (#2253)
