+++ src/synthorg/api/lifecycle_helpers/config_apply.py
@@ -1,3 +1,4 @@
+# module-kind: orchestrator
 """Apply operator-tuned bridge config at startup.
 
 Snapshots ``ApiBridgeConfig`` onto ``AppState``, validates cross-
